+++ apps/frontend/src/pages/access-management/access-management.tsx
@@ -178,12 +178,15 @@ const AccessManagement = (): JSX.Element => {
 		handleLoadGroups,
 	]);
 
-	const isLoading = [usersDataStatus, groupsDataStatus].some(
-		(status) => status === DataStatus.IDLE || status === DataStatus.PENDING,
-	);
+	const isLoadingUsersData =
+		usersDataStatus === DataStatus.IDLE ||
+		usersDataStatus === DataStatus.PENDING;
+	const isLoadingGroupsData =
+		groupsDataStatus === DataStatus.IDLE ||
+		groupsDataStatus === DataStatus.PENDING;
 
 	return (
-		<PageLayout isLoading={isLoading}>
+		<PageLayout>
 			<h1 className={styles["title"]}>Access Management</h1>
 
 			<section>
@@ -191,6 +194,7 @@ const AccessManagement = (): JSX.Element => {
 					<h2 className={styles["section-title"]}>Users</h2>
 				</div>
 				<UsersTable
+					isLoading={isLoadingUsersData}
 					onPageChange={onUserPageChange}
 					onPageSizeChange={onUserPageSizeChange}
 					page={userPage}
@@ -210,6 +214,7 @@ const AccessManagement = (): JSX.Element => {
 
 				<GroupsTable
 					groups={groups}
+					isLoading={isLoadingGroupsData}
 					onDelete={handleDelete}
 					onEdit={handleEdit}
 					onPageChange={onGroupPageChange}