+++ packages/shared/src/index.ts
@@ -58,6 +58,7 @@ export {
 	ContributorError,
 	type ContributorGetAllItemResponseDto,
 	type ContributorGetAllResponseDto,
+	ContributorsApiPath,
 } from "./modules/contributors/contributors.js";
 export {
 	type GitEmailCreateRequestDto,
@@ -104,8 +105,16 @@ export {
 	type ProjectGroupCreateResponseDto,
 	projectGroupCreateValidationSchema,
 	ProjectGroupError,
+	type ProjectGroupGetAllItemResponseDto,
+	type ProjectGroupGetAllRequestDto,
+	type ProjectGroupGetAllResponseDto,
 	ProjectGroupsApiPath,
 } from "./modules/project-groups/project-groups.js";
+export {
+	ProjectPermissionsApiPath,
+	type ProjectPermissionsGetAllItemResponseDto,
+	type ProjectPermissionsGetAllResponseDto,
+} from "./modules/project-permissions/project-permissions.js";
 export {
 	type ProjectCreateRequestDto,
 	projectCreateValidationSchema,