+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -1,11 +1,13 @@
 const APIPath = {
 	ACTIVITY_LOGS: "/activity-logs",
 	AUTH: "/auth",
+	CONTRIBUTORS: "/contributors",
 	GROUPS: "/groups",
 	NOTIFICATIONS: "/notifications",
 	PERMISSIONS: "/permissions",
 	PROJECT_API_KEYS: "/project-api-keys",
 	PROJECT_GROUPS: "/project-groups",
+	PROJECT_PERMISSIONS: "/project-permissions",
 	PROJECTS: "/projects",
 	USERS: "/users",
 } as const;