+++ packages/shared/src/modules/contributors/contributors.ts
@@ -1,3 +1,4 @@
+export { ContributorsApiPath } from "./libs/enums/enums.js";
 export { ContributorError } from "./libs/exceptions/exceptions.js";
 export {
 	type ContributorCreateRequestDto,