+++ packages/shared/src/modules/project-groups/libs/types/project-group-create-response-dto.type.ts
@@ -1,9 +1,10 @@
 type ProjectGroupCreateResponseDto = {
+	createdAt: string;
 	id: number;
 	name: string;
-	permissions: { id: number }[];
+	permissions: { id: number; name: string }[];
 	projectId: { id: number };
-	users: { id: number }[];
+	users: { createdAt: string; email: string; id: number; name: string }[];
 };
 
 export { type ProjectGroupCreateResponseDto };